+++ src/Infrastructure/Bundle/InfrastructureBundle/Validator/Constraint/Password/ContainsAtLeastOneLowercaseValidator.php
@@ -9,6 +9,10 @@ class ContainsAtLeastOneLowercaseValidator extends ConstraintValidator
 {
     public function validate($value, Constraint $constraint)
     {
+        if (null === $value) {
+            return;
+        }
+
         if (!preg_match("~[a-z]~", $value)) {
             $this->context->buildViolation($constraint->message)
                 ->addViolation();